requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
supabase>=2.0
openai>=1.0
pydantic>=2.0
python-dotenv>=1.0
